requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==4.2.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.28.40
botocore==1.31.40
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
filelock==3.12.2
Flask==2.2.5
Flask-Sessionstore==0.4.5
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.10
GitPython==3.1.31
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.6.0
install==1.3.5
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.5
lightning-utilities==0.9.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==2.6.3
numpy==1.21.1
packaging==23.1
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
protobuf==3.20.3
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytorch-lightning==1.9.5
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
requests==2.28.2
rich==13.3.4
# rpy2==3.5.13
s3transfer==0.6.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
stack-data==0.6.2
streamlit==1.21.0
sympy==1.10.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.2.2
WTForms==3.0.1
yarl==1.9.2
zipp==3.15.0
